requirements.txt
streamlit
python-docx
faster-whisper
torch
torchaudio
torchvision
transformers
ffmpeg-python
librosa
scikit-learn
numpy
python-speech-features
watchdog
